requirements.txt
# Core LangGraph and LLM
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# Document Processing
python-docx>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-magic>=0.4.27
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Graph and Data Structures
networkx>=3.0
pydantic>=2.0.0
pandas>=2.0.0

# NLP and Pattern Matching
spacy>=3.7.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Visualization
# graphviz>=0.20.0
graphviz
pyvis
matplotlib>=3.7.0
plotly>=5.17.0
reportlab>=4.0.0
weasyprint>=60.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
